Abstract csv, issues fns to their own files; add issues.test.ts; add tsdown and optional deps for build step
