+++ package.json
@@ -6,18 +6,16 @@
   "types": "dist/index.d.ts",
   "exports": {
     ".": {
-      "import": "./dist/index.js",
-      "types": "./dist/index.d.ts"
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
     }
   },
   "files": [
     "dist"
   ],
   "scripts": {
-    "build": "tsc",
-    "dev": "tsc --watch",
-    "clean": "rm -rf dist",
-    "prebuild": "npm run clean",
+    "build": "tsdown",
+    "dev": "tsdown --watch",
     "test": "vitest run",
     "test:watch": "vitest",
     "publish": "npm run build && npm publish --access public"
@@ -27,6 +25,9 @@
   "license": "ISC",
   "description": "",
   "devDependencies": {
+    "@arethetypeswrong/core": "^0.18.2",
+    "publint": "^0.3.18",
+    "tsdown": "^0.21.7",
     "typescript": "^5.8.3",
     "vitest": "^4.1.4"
   },
@@ -35,7 +36,7 @@
     "zod": "^4.0.5"
   },
   "engines": {
-    "node": ">=24.14.1",
-    "npm": ">=11.11.0"
+    "node": "^24",
+    "npm": "^11"
   }
 }
