+++ tsdown.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from 'tsdown';
+
+export default defineConfig({
+  attw: {
+    profile: "esm-only"
+  },
+  platform: "neutral",
+  publint: true,
+});
